+++ src/simmate/apps/quantum_espresso/command_line/base_command.py
@@ -17,11 +17,9 @@ qe_app = typer.Typer(
 @qe_app.callback(no_args_is_help=True)
 def base_command():
     """
-    A collection of utilities to help work with VASP
+    Commands and utilities for managing Quantum Espresso (QE) calculations
+    and configurations.
     """
-    # When we call the command "simmate-vasp" this is where we start, and it then
-    # looks for all other functions that have the decorator "@simmate_vasp.command()"
-    # to decide what to do from there.
     pass