+++ src/simmate/command_line/test/test_worker_cli.py
@@ -2,14 +2,14 @@
 
 import pytest
 
-from simmate.command_line.engine import engine_app
+from simmate.command_line.compute import compute_app
 
 
 @pytest.mark.django_db
 def test_start_worker_cli(command_line_runner):
     # dump the database to json
     result = command_line_runner.invoke(
-        engine_app,
+        compute_app,
         ["start-singleflow-worker"],
     )
 
@@ -17,7 +17,7 @@ def test_start_worker_cli(command_line_runner):
 
     # load the database to json
     result = command_line_runner.invoke(
-        engine_app,
+        compute_app,
         ["start-worker", "--nitems-max", "1", "--close-on-empty-queue"],
     )
     assert result.exit_code == 0